requirements.txt
python-dotenv
pandas
numpy
fastapi==0.104.1
uvicorn==0.24.0
python-binance==1.0.16
websockets==11.0.3
aiohttp==3.9.1